+++ docroot/WEB-INF/sql/tables.sql
@@ -1,5 +1,5 @@
 create table StaffOrderStuff_Consumption (
-	ConsumptionId INTEGER not null primary key,
+	ConsumptionId LONG not null primary key,
 	UserId INTEGER,
 	FoodId INTEGER,
 	Date DATE null,
@@ -9,7 +9,7 @@ create table StaffOrderStuff_Consumption (
 );
 
 create table StaffOrderStuff_Food (
-	FoodId INTEGER not null primary key,
+	FoodId LONG not null primary key,
 	CompanyId LONG,
 	GroupId LONG,
 	Name VARCHAR(75) null,
